+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/TrimmingAudioProcessor.java
@@ -19,7 +19,6 @@ import static androidx.media3.common.util.Util.isEncodingLinearPcm;
 import static java.lang.Math.max;
 import static java.lang.Math.min;
 
-import androidx.media3.common.Format;
 import androidx.media3.common.audio.BaseAudioProcessor;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
@@ -48,7 +47,7 @@ public final class TrimmingAudioProcessor extends BaseAudioProcessor {
    * processor. After calling this method, call {@link #configure(AudioFormat)} to apply the new
    * trimming frame counts.
    *
-   * <p>See {@link AudioSink#configure(Format, int, int[])}.
+   * <p>See {@link AudioSink#configure(androidx.media3.exoplayer.audio.AudioSink.AudioSinkConfig)}.
    *
    * @param trimStartFrames The number of audio frames to trim from the start of audio.
    * @param trimEndFrames The number of audio frames to trim from the end of audio.