+++ RELEASENOTES.md
@@ -22,6 +22,9 @@
 *   Extractors:
 *   Inspector:
 *   Audio:
+    *   Convert parameters of `AudioSink.configure` to data class. Custom
+        overrides of `ForwardingAudioSink.configure` need to be migrated to the
+        new method signature.
 *   Video:
     *   Add support for skipping frames that are late during join rather than
         dropping in DecoderVideoRenderer.